+++ src/fd_time.c
@@ -47,6 +47,9 @@ fd_time_t fd_time_multiply(fd_time_t t, unsigned n) {
     uint64_t us = ((uint64_t)t.microseconds) * n;
     uint64_t s = us / 1000000;
     us = us - (s * 1000000);
-    fd_time_t r = {.seconds = s, .microseconds = us};
+    fd_time_t r = {
+        .seconds = (uint32_t)s,
+        .microseconds = (uint32_t)us
+    };
     return r;
 }
